Refactor trpc admin routes, better type safety for permissions
