+++ apps/web/src/components/auth/permission/client/gate.tsx
@@ -5,7 +5,7 @@ import { useRouter, usePathname } from "next/navigation";
 import { PermissionIdentifier } from "@repo/db/client";
 import { usePermissions } from "../utils/usePermissions";
 import { isPublicPath } from "../utils/path-utils";
-import { logger } from "~/lib/utils/logger";
+import { logger } from "@repo/logger";
 
 // Define prop types for each slot component
 interface AuthorizedProps {
