+++ apps/web/src/lib/markdown/plugins/server-only/loggerPlugin.ts
@@ -5,7 +5,7 @@
 
 import { visit } from "unist-util-visit";
 import type { Plugin } from "unified";
-import { logger } from "~/lib/utils/logger";
+import { logger } from "@repo/logger";
 
 interface LoggerPluginOptions {
   /**
