+++ apps/web/src/server/routers/index.ts
@@ -1,22 +1,18 @@
 import { router } from "..";
 import { wikiRouter } from "./wiki";
-import { userRouter } from "./user";
+import { usersRouter } from "./users";
 import { searchRouter } from "./search";
 import { assetsRouter } from "./assets";
-import { permissionsRouter } from "./permissions";
-import { groupsRouter } from "./groups";
-import { usersRouter } from "./users";
 import { authRouter } from "./auth";
 import { tagsRouter } from "./tags";
+import { adminRouter } from "./admin";
 
 export const appRouter = router({
+  admin: adminRouter,
   wiki: wikiRouter,
-  user: userRouter,
   users: usersRouter,
   search: searchRouter,
   assets: assetsRouter,
-  permissions: permissionsRouter,
-  groups: groupsRouter,
   auth: authRouter,
   tags: tagsRouter,
 });
