+++ packages/db/package.json
@@ -36,7 +36,9 @@
   },
   "dependencies": {
     "@neondatabase/serverless": "^1.0.0",
+    "@repo/logger": "workspace:*",
     "bcryptjs": "^3.0.2",
+    "child_process": "^1.0.2",
     "dotenv": "^16.5.0",
     "drizzle-orm": "^0.42.0",
     "gray-matter": "^4.0.3",
