+++ packages/db/src/registry/types.ts
@@ -35,6 +35,9 @@ export interface Permission {
 
 /**
  * Type for the permission identifier string in format module:resource:action
+ * Represents any possible combination based on defined modules, resources, and actions.
+ * For a stricter type derived from actual registered permissions, use `PermissionIdentifier`
+ * from `packages/db/src/registry/index.ts`.
  */
-export type PermissionIdentifier =
+export type PossiblePermissionIdentifier =
   `${PermissionModule}:${PermissionResource}:${PermissionAction}`;
